Leverage and export per-frame AHardwareBuffer objects
Instead of collecting and rendering against an ANativeWindow object, manually
spawn and use AHardwareBuffer objects as framebuffer backings into which the
Web content is composited. Each such object is then exported to the
corresponding wpe_view_backend object, leaving it to the embedder to handle
the AHardwareBuffer as appropriate.

For each view, a pool of these objects is created and managed between the
view backend and the renderer backend. Rendering into an unlocked buffer
eventually locks it and leaves it to the view backend to export the buffer
to the embedder, until the embedder returns it and the buffer gets unlocked
and available for any following frame.

One optional change, for future assessment, is managing the buffer pool from
the view backend side, instead of from the renderer backend.